Fixed template dir and added drag and drop support
