--- .nvmrc.old
@@ -1,1 +1,0 @@
-lts/hydrogen
