+++ src/app.css
@@ -54,11 +54,13 @@ a:visited {
 /* Container Styles */
 .container {
   flex: 1;
-  margin: 0 auto; /* Center the container */
+  margin: 0 auto;
   display: flex;
   flex-direction: column;
   align-items: center;
   justify-content: center;
+  width: clamp(300px, 50%, 1000px);
+  box-sizing: border-box;
 }
 
 /* Card Styles */
@@ -68,21 +70,16 @@ a:visited {
   color: var(--text-color);
   border-radius: 8px;
   padding: 30px;
-  margin: auto;
-  min-width: 300px;
-  max-width: 400px;
-}
-
-/* File Input Styles */
-input[type='file'] {
-  background-color: var(--background-color);
-  color: var(--text-color);
-  padding: 10px;
-  border-radius: 5px;
+  width: 100%;
+  max-width: 960px;
+  margin: 0 auto;
+  box-sizing: border-box;
+  text-align: center;
 }
 
 /* Button Styles */
 button {
+  display: inline-block;
   background-color: var(--primary-color);
   color: #fff;
   padding: 10px 20px;
@@ -91,7 +88,9 @@ button {
   cursor: pointer;
   font-size: 16px;
   width: 100%;
-  margin: 3em auto 1em;
+  max-width: 350px;
+  margin: 25px auto;
+  box-sizing: border-box;
 }
 
 button:hover {
@@ -99,14 +98,66 @@ button:hover {
 }
 
 /* File Input Button Styles */
+.drop-container {
+  position: relative;
+  display: flex;
+  gap: 10px;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  height: 350px;
+  padding: 20px;
+  border-radius: 10px;
+  border: 2px dashed #555;
+  color: var(--text-color)
+  cursor: pointer;
+  transition:
+    background 0.2s ease-in-out,
+    border 0.2s ease-in-out;
+}
+
+.drop-title {
+  color: var(--text-color);
+  font-size: 20px;
+  font-weight: bold;
+  text-align: center;
+  transition: color 0.2s ease-in-out;
+}
+
+.drop-container:hover,
+.drop-container.drag-active {
+  background: #eee;
+  border-color: #111;
+}
+
+.drop-container:hover .drop-title,
+.drop-container.drag-active .drop-title {
+  color: #222;
+}
+
+input[type='file'] {
+  width: 350px;
+  max-width: 100%;
+  color: #444;
+  padding: 5px;
+  background: #fff;
+  border-radius: 10px;
+  border: 1px solid #555;
+}
+
 input[type='file']::file-selector-button {
-  padding: 0.3em 0.5em;
-  border-radius: 0.3em;
-  margin: 5px;
+  margin-right: 20px;
+  border: none;
+  background: #084cdf;
+  padding: 10px 20px;
+  border-radius: 10px;
+  color: #fff;
+  cursor: pointer;
+  transition: background 0.2s ease-in-out;
 }
 
 input[type='file']::file-selector-button:hover {
-  background-color: var(--secondary-background-color);
+  background: #0d45a5;
 }
 
 /* Alert Styles */
@@ -136,38 +187,4 @@ footer {
   padding: 20px;
   margin-top: auto; /* Push the footer to the bottom */
   font-size: small;
-}
-
-/* Media queries for common screen sizes */
-@media only screen and (width < 600px) {
-  /* small screens */
-  .container {
-    width: 100%;
-  }
-
-  .card {
-    width: 90%;
-  }
-}
-
-@media only screen and (width >= 601px) and (width <= 1024px) {
-  /* medium screens */
-  .container {
-    width: 80%;
-  }
-
-  .card {
-    width: 100%;
-  }
-}
-
-@media only screen and (width >= 1025px) {
-  /* large screens */
-  .container {
-    width: 50%;
-  }
-
-  .card {
-    width: 100%;
-  }
 }
