+++ src/config.ts
@@ -2,7 +2,7 @@ const config = {
   paths: {
     public: '/public',
     data: '/data',
-    emTemplate: '/public/templates/em.xlsx',
+    emTemplate: 'templates/em.xlsx',
     beTemplate: 'BLD-BE-001 Budget Estimate template.xlsx',
   },
 };
